Fix sign out button pending state; refactor folder structure
